# openapi_backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClaimVision-API",
    "description": "API del proyecto ClaimVision, en constante actualizacion y mejoria",
    "version": "1.12.0"
  },
  "paths": {
...
    "/api/v1/taller/perfil": {
      "get": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Get Perfil",
        "description": "§6 · Perfil del taller (datos del taller + datos del operador).",
        "operationId": "get_perfil_api_v1_taller_perfil_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerPerfilResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Actualizar Perfil",
        "description": "§6 · Actualiza perfil del taller (datos del taller + datos del operador).",
        "operationId": "actualizar_perfil_api_v1_taller_perfil_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TallerPerfilUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerPerfilResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/taller/ordenes": {
      "get": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Listar Ordenes",
        "description": "§6 · Órdenes del taller (Asignado_A_Taller / en proceso).",
        "operationId": "listar_ordenes_api_v1_taller_ordenes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "estatus",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SiniestroResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/taller/siniestros/{id}": {
      "get": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Detalle Orden",
        "description": "§6 · Expediente técnico (siniestro + peritaje validado + cotización).",
        "operationId": "detalle_orden_api_v1_taller_siniestros__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TallerExpedienteDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/taller/siniestros/{id}/cotizacion": {
      "post": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Crear Cotizacion",
        "description": "§6 · Crea cotización → estatus_cotizacion = Pendiente_Aprobacion. Recibe PDF del desglose.",
        "operationId": "crear_cotizacion_api_v1_taller_siniestros__id__cotizacion_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_crear_cotizacion_api_v1_taller_siniestros__id__cotizacion_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CotizacionV1DTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/taller/cotizaciones/{id}": {
      "patch": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Editar Cotizacion",
        "description": "§6 · Edita cotización mientras siga Pendiente_Aprobacion.",
        "operationId": "editar_cotizacion_api_v1_taller_cotizaciones__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarCotizacionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CotizacionV1DTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/taller/siniestros/{id}/concluir-trabajo": {
      "post": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Concluir Trabajo",
        "description": "§6 · → estatus = Trabajo_Concluido (requiere cotización aprobada).",
        "operationId": "concluir_trabajo_api_v1_taller_siniestros__id__concluir_trabajo_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/taller/siniestros/{id}/listo-entrega": {
      "post": {
        "tags": [
          "v1 · Taller"
        ],
        "summary": "Listo Entrega",
        "description": "§6 · → estatus = Listo_Para_Entrega.",
        "operationId": "listo_entrega_api_v1_taller_siniestros__id__listo_entrega_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiniestroResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_crear_cotizacion_api_v1_taller_siniestros__id__cotizacion_post": {
        "properties": {
          "monto_mano_obra": {
            "type": "number",
            "title": "Monto Mano Obra"
          },
          "monto_refacciones": {
            "type": "number",
            "title": "Monto Refacciones"
          },
          "monto_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monto Total"
          },
          "observaciones_tecnicas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones Tecnicas"
          },
          "desglose_pdf": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Desglose Pdf"
          }
        },
        "type": "object",
        "required": [
          "monto_mano_obra",
          "monto_refacciones",
          "desglose_pdf"
        ],
        "title": "Body_crear_cotizacion_api_v1_taller_siniestros__id__cotizacion_post"
      },
...
      "CotizacionV1DTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "siniestro_id": {
            "type": "string",
            "title": "Siniestro Id"
          },
          "taller_id": {
            "type": "string",
            "title": "Taller Id"
          },
          "monto_mano_obra": {
            "type": "number",
            "title": "Monto Mano Obra"
          },
          "monto_refacciones": {
            "type": "number",
            "title": "Monto Refacciones"
          },
          "monto_total": {
            "type": "number",
            "title": "Monto Total"
          },
          "desglose_pdf_url": {
            "type": "string",
            "title": "Desglose Pdf Url"
          },
          "estatus": {
            "type": "string",
            "title": "Estatus"
          },
          "observaciones_tecnicas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones Tecnicas"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "siniestro_id",
          "taller_id",
          "monto_mano_obra",
          "monto_refacciones",
          "monto_total",
          "desglose_pdf_url",
          "estatus",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "CotizacionV1DTO"
      },
...
      "EditarCotizacionRequest": {
        "properties": {
          "monto_mano_obra": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monto Mano Obra"
          },
          "monto_refacciones": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monto Refacciones"
          },
          "monto_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monto Total"
          },
          "desglose_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desglose Pdf Url"
          },
          "observaciones_tecnicas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones Tecnicas"
          }
        },
        "type": "object",
        "title": "EditarCotizacionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_SiniestroResponseDTO_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SiniestroResponseDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[SiniestroResponseDTO]"
      },
...
      "TallerExpedienteDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "aseguradora_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aseguradora Id"
          },
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "ajustador_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ajustador Id"
          },
          "taller_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taller Id"
          },
          "vehiculo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiculo Id"
          },
          "estatus": {
            "type": "string",
            "title": "Estatus"
          },
          "vehiculo_marca": {
            "type": "string",
            "title": "Vehiculo Marca"
          },
          "vehiculo_modelo": {
            "type": "string",
            "title": "Vehiculo Modelo"
          },
          "vehiculo_anio": {
            "type": "integer",
            "title": "Vehiculo Anio"
          },
          "vehiculo_placas": {
            "type": "string",
            "title": "Vehiculo Placas"
          },
          "vehiculo_vin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiculo Vin"
          },
          "latitud_siniestro": {
            "type": "number",
            "title": "Latitud Siniestro"
          },
          "longitud_siniestro": {
            "type": "number",
            "title": "Longitud Siniestro"
          },
          "narracion_texto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narracion Texto"
          },
          "narracion_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narracion Audio Url"
          },
          "indicaciones_dano_interno": {
            "type": "boolean",
            "title": "Indicaciones Dano Interno"
          },
          "fecha_siniestro": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Siniestro"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "peritaje": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PeritajeResponseDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "cotizacion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CotizacionV1DTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "cliente_id",
          "estatus",
          "vehiculo_marca",
          "vehiculo_modelo",
          "vehiculo_anio",
          "vehiculo_placas",
          "latitud_siniestro",
          "longitud_siniestro",
          "indicaciones_dano_interno",
          "fecha_siniestro",
          "created_at"
        ],
        "title": "TallerExpedienteDTO"
      },
      "TallerPerfilResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nombre_comercial": {
            "type": "string",
            "title": "Nombre Comercial"
          },
          "rfc": {
            "type": "string",
            "title": "Rfc"
          },
          "direccion_tecnica": {
            "type": "string",
            "title": "Direccion Tecnica"
          },
          "telefono_contacto": {
            "type": "string",
            "title": "Telefono Contacto"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre_comercial",
          "rfc",
          "direccion_tecnica",
          "telefono_contacto",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "TallerPerfilResponse"
      },
      "TallerPerfilUpdateRequest": {
        "properties": {
          "nombre_comercial": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Comercial"
          },
          "direccion_tecnica": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direccion Tecnica"
          },
          "telefono_contacto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono Contacto"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          }
        },
        "type": "object",
        "title": "TallerPerfilUpdateRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}